+++ src/app/layout.tsx
@@ -1,4 +1,4 @@
-import type { Metadata } from "next";
+import type { Metadata, Viewport } from "next";
 import { Geist, Geist_Mono, Inter } from "next/font/google";
 import "./globals.css";
 import "katex/dist/katex.min.css"
@@ -64,6 +64,10 @@ export const metadata: Metadata = {
     }]
   },
 };
+
+export const viewport: Viewport = {
+  themeColor: "#0069a8"
+}
 
 export default function RootLayout({ children }: LayoutProps<"/">) {
   return (