+++ src/app/robots.ts
@@ -1,0 +1,14 @@
+import { absoluteURL } from "@/lib/utils";
+import type { MetadataRoute } from "next";
+
+export default function robots(): MetadataRoute.Robots {
+     return {
+          rules: [
+               {
+                    userAgent: "*",
+                    allow: "/",
+               },
+          ],
+          sitemap: absoluteURL("/sitemap.xml"),
+     };
+}