+++ src/lib/helpers/templates.ts
@@ -1,16 +1,13 @@
 import type { Template } from "@/lib/types";
-import { absoluteCDN, getErrorMessage } from "../utils"
+import { absoluteCDN } from "../utils"
 
-export const getTemplatesFromCDN = async (): Promise<{
-     data: Template[],
-     error?: string
-}> => {
+export const getTemplatesFromCDN = async (): Promise<Template[]> => {
      try {
           const res = await fetch(absoluteCDN("/templates.json"));
-          if(!res.ok) return {error: "Չհաջողվեց բեռնել շաբլոնների ցուցակը", data: []}
+          if(!res.ok) return []
           const data: Template[] = await res.json();
-          return {data}
-     } catch (error) {
-          return {error: getErrorMessage(error), data: []}
+          return data
+     } catch {
+          return []
      }
 }