+++ src/app/experiments/[slug]/page.tsx
@@ -2,7 +2,7 @@ import { TooltipProvider } from "@/components/ui/tooltip";
 import SingleExperimentContent from "@/contents/single-experiment";
 import { KEYWORDS } from "@/lib/constants";
 import { getAllSlugs, getExperimentBySlug, getRelatedExperiments } from "@/lib/helpers/experiments";
-import { createMetaAlternates } from "@/lib/utils";
+import { absoluteURL, createMetaAlternates } from "@/lib/utils";
 import { Metadata } from "next";
 import { notFound } from "next/navigation";
 
@@ -20,6 +20,11 @@ export const generateMetadata = async({params}:SingleExperimentPageProps): Promi
      const {slug} = await params;
      const currExperiment = await getExperimentBySlug(slug);
      if(!currExperiment) return notFound();
+     const ogSearchParams = new URLSearchParams({
+          title: currExperiment.title,
+          thumbnail: currExperiment.thumbnail,
+          date: currExperiment.date
+     })
      return {
           title: currExperiment.title,
           description: currExperiment.description,
@@ -30,6 +35,28 @@ export const generateMetadata = async({params}:SingleExperimentPageProps): Promi
                     url: "https://arsen-2005.vercel.app/"
                }
           ],
+          openGraph: {
+               title: currExperiment.title,
+               description: currExperiment.description,
+               url: absoluteURL(`/experiments/${slug}`),
+               siteName: "Գիտաֆիզիկա",
+               type: "article",
+               images: {
+                    url: absoluteURL(`/api/og?${ogSearchParams.toString()}`),
+                    width: 1200,
+                    height: 630
+               }
+          },
+          twitter: {
+               title: currExperiment.title,
+               description: currExperiment.description,
+               card: "summary_large_image",
+               images: [{
+                    url: absoluteURL(`/api/og?${ogSearchParams.toString()}`),
+                    width: 1200,
+                    height: 630
+               }]
+          },
           alternates: createMetaAlternates(`/experiments/${slug}`)
      }
 }